backend/requirements.txt
google-genai
websockets
pydantic
python-dotenv
colorama
